fixup! test(studio): add tests for data api toggles
